swapped test files locations to resemble app structure
