+++ src/tests/forms/RegisterPatientForm.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import {render, screen, waitFor} from "@testing-library/react";
 import user from '@testing-library/user-event'
-import RegisterPatientForm from "../components/RegisterPatientForm";
+import RegisterPatientForm from "../../components/forms/RegisterPatientForm";
 
 
 describe('RegisterPatientForm', () => {
