+++ src/tests/pages/DoctorDashboard.test.tsx
@@ -1,0 +1,11 @@
+import {render, screen, waitFor} from "@testing-library/react";
+import user from "@testing-library/user-event";
+
+describe("DoctorDashboard", () => {
+    it('routes to setSchedule', async () => {
+        //TODO
+    });
+    it('calls API for visits', async () => {
+        //TODO
+    });
+});
